+++ brewtils/schemas.py
@@ -477,8 +477,6 @@ class ConnectionSchema(BaseSchema):
 class GardenSchema(BaseSchema):
     id = fields.Str(allow_none=True)
     name = fields.Str(allow_none=True)
-    status = fields.Str(allow_none=True)
-    status_info = fields.Nested("StatusInfoSchema", allow_none=True)
     connection_type = fields.Str(allow_none=True)
     receiving_connections = fields.Nested(
         "ConnectionSchema", many=True, allow_none=True
@@ -486,7 +484,6 @@ class GardenSchema(BaseSchema):
     publishing_connections = fields.Nested(
         "ConnectionSchema", many=True, allow_none=True
     )
-    namespaces = fields.List(fields.Str(), allow_none=True)
     systems = fields.Nested("SystemSchema", many=True, allow_none=True)
     has_upstream = fields.Bool(allow_none=True)
     upstream = fields.Str(allow_none=True)